Opitional task

+++ src/main/java/vangoh74/booklibrary/service/BookService.java
@@ -4,6 +4,8 @@ import vangoh74.booklibrary.model.Book;
 import vangoh74.booklibrary.repository.BookRepository;
 
 import java.util.List;
+import java.util.NoSuchElementException;
+import java.util.Optional;
 
 @Service
 public class BookService {
@@ -22,10 +24,23 @@ public class BookService {
     }
 
     public Book getBookByISBN(String isbn) {
-        return bookRepo.getBookByISBN(isbn);
+        Optional<Book> optBook = bookRepo.getBookByISBN(isbn);
+        if (optBook.isEmpty()) {
+            throw new NoSuchElementException();
+        }
+        return optBook.get();
+        //return bookRepo.getBookByISBN(isbn)
+            //    .orElseThrow(() -> new NoSuchElementException("Book with isbn " + isbn + " is not found!"));
     }
 
     public Book deleteBook(String isbn) {
-        return bookRepo.deleteBook(isbn);
+       /*
+        Optional<Book> optBook = bookRepo.deleteBook(isbn);
+        if (optBook.isEmpty()) {
+            throw new NoSuchElementException();
+        }
+        return optBook.get();*/
+        return bookRepo.deleteBook(isbn)
+                .orElseThrow(() -> new NoSuchElementException("Book with isbn " + isbn + " is not found!"));
     }
 }
